refatorando codigos

+++ src/main/java/com/misticazoo/misticazoo/dto/ItemRequestDTO.java
@@ -1,4 +1,4 @@
 package com.misticazoo.misticazoo.dto;
 
-public record ItemRequestDTO(String nome, String descricao, Integer preco, Integer estoque) {
+public record ItemRequestDTO(String nome, String descricao, double preco, Integer estoque) {
 }
