+++ src/main/java/com/misticazoo/misticazoo/dto/PagamentoRequestDTO.java
@@ -1,4 +1,4 @@
 package com.misticazoo.misticazoo.dto;
 
-public record PagamentoRequestDTO(String dataPagamento, Integer valor, String metodo, String status) {
+public record PagamentoRequestDTO(String dataPagamento, double valor, String metodo, String status) {
 }
