+++ src/main/java/com/misticazoo/misticazoo/dto/PedidoRequestDTO.java
@@ -1,4 +1,4 @@
 package com.misticazoo.misticazoo.dto;
 
-public record PedidoRequestDTO(String dataPedido, String status, Integer valorTotal) {
+public record PedidoRequestDTO(String dataPedido, String status, double valorTotal) {
 }
